chore: Heic 파일을 WebP 파일로 변환하는 함수를 공통 함수로 파일 분리
